Introduce Google Speech-To-Text V2 library
